requirements.txt
streamlit>=1.36
pandas>=2.0
plotly>=5.18
mysql-connector-python>=8.0
